% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_gene_lengths.R
\docType{data}
\name{ensembl_gene_lengths}
\alias{ensembl_gene_lengths}
\title{Gene Lengths from the Ensembl Database}
\format{
A dataframe with two columns:
\describe{
\item{Hugo_Symbol}{The names of all nuclear genes in humans for which ensembl entries with coding sequence lengths exist.}
\item{max_cds}{The maximum coding sequence for each gene as given by the ensembl database.}
}
}
\source{
\url{https://www.ensembl.org/}
}
\usage{
ensembl_gene_lengths
}
\description{
Gene Lengths from the Ensembl Database
}
\keyword{datasets}
