% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_first_pred_tmb}
\alias{example_first_pred_tmb}
\title{First-Fit Predictive Model Fitting on Example Data}
\format{
A list with six entries:
\describe{
\item{fit}{A gglasso fit.}
\item{panel_genes}{A matrix where each row corresponds to a gene, each column to an iteration of the group
lasso with a different penalty factor, and the elements booleans specifying whether that gene was selected
to be included in that iteration.}
\item{panel_lengths}{A vector giving total panel length for each gglasso iteration.}
\item{p}{The vector of weights used in the optimisation procedure.}
\item{K}{The bias penalty factor used in the optimisation procedure.}
\item{names}{Gene and mutation type information as used when fitting the generative model.}

}
}
\usage{
example_first_pred_tmb
}
\description{
An example output from the function pred_first_fit(), applied to pre-loaded example mutation data.
}
\keyword{datasets}
