% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_gen_model}
\alias{example_gen_model}
\title{Generative Model from Simulated Data}
\format{
A list with two entries:
\describe{
\item{fit}{A glmnet fit object.}
\item{dev}{A table containing the average deviance of each cross-validation fold, for each penalisation factor in fit$lambda.}
\item{s_min}{The index of the regularisation penalty minimising average deviance across folds.}
}
}
\usage{
example_gen_model
}
\description{
An example of the output produced by fit_gen_model() on simulated data.
}
\keyword{datasets}
