% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_maf_data}
\alias{example_maf_data}
\title{Simulated MAF Data}
\format{
A list with two entries:
\describe{
\item{maf}{An annotated mutation dataframe with 3 columns and 1346 rows: \describe{
\item{Tumor_Sample_Barcode}{A sample id for each mutation.}
\item{Hugo_Symbol}{The name of the gene location for each mutation.}
\item{Variant_Classification}{The mutation type for each mutation.}}}
\item{gene_lengths}{A data frame with two rows: \describe{
\item{Hugo_Symbol}{The name of each gene.}
\item{max_cds}{The length of each gene, as defined by maximum coding sequence.}}}
}
}
\usage{
example_maf_data
}
\description{
An example dataset generated by the function generate_maf_data(), with n_sample = 100 and n_genes = 20.
}
\keyword{datasets}
