% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_predictions}
\alias{example_predictions}
\title{Example Predictions}
\format{
A list with two entries:
\describe{
\item{predictions}{A a matrix containing a row for each sample and a column for each panel.}
\item{panel_lengths}{A vector giving total panel lengths.}
}
}
\usage{
example_predictions
}
\description{
An example output from use of the function get_predictions(), applied to the pre-loaded datasets example_refit_range and example_tables$val .
}
\keyword{datasets}
