% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_refit_range}
\alias{example_refit_range}
\title{Refitted Predictive Models Fitted on Example Data}
\format{
A list with six entries:
\describe{
\item{fit}{A list with a single element 'beta', a matrix with prediction weights.}
\item{panel_genes}{A matrix where each row corresponds to a gene,
and each entry corresponds to whether the gene is included in the panel.}
\item{panel_lengths}{A vector giving total panel length.}
}
}
\usage{
example_refit_range
}
\description{
An example output from use of the function pred_refit_range(), applied to example gene length data and generative model fit.
}
\keyword{datasets}
