% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_tables}
\alias{example_tables}
\title{Simulated MAF Data}
\format{
A list with three entries:
\describe{
\item{train}{An object 'train'.}
\item{val}{An object 'val'.}
\item{test}{An object 'test'.}
}
Each of these three objects is a list with the following entries (for more detail see the documentation for the function get_table_from_maf()):
\describe{
\item{matrix}{A sparse matrix of mutations.}
\item{sample_list}{A character vector of sample IDs, corresponding to the rows of the mutation matrix.}
\item{gene_list}{A characer vector of gene names.}
\item{mut_types_list}{A character vector of mutation types.}
\item{colnames}{A character vector of gene name/mutation type combinations (in each case separated by the character "_"), corresponding to the columns of the mutation matrix.}
}
}
\usage{
example_tables
}
\description{
An example dataset generated by the function generate_maf_data(), with n_sample = 20 and n_genes = 50.
}
\keyword{datasets}
