% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_tmb_tables}
\alias{example_tmb_tables}
\title{Tumour Mutation Burden of Example Train, Validation and Test Data.}
\format{
A list with thre objects: 'train', 'val' and 'test'. Each is a dataframe with two columns:
\describe{
\item{Tumor_Sample_Barcode}{A unique ID for each sample.}
\item{TMB}{The value of Tumour Mutation Burden for that sample.}
}
}
\usage{
example_tmb_tables
}
\description{
Produced using the function get_biomarker_tables(), and from the example MAF data attached in example_maf_data$maf.
}
\keyword{datasets}
