% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICBioMark.R
\docType{package}
\name{ICBioMark}
\alias{ICBioMark}
\title{ICBioMark: A package for cost-effective design of gene panels to predict
exome-wide biomarkers.}
\description{
This package implements the methodology proposed in
'Data-driven design of targeted gene panels for estimating immunotherapy biomarkers',
(Bradley and Cannings, 2021, preprint). It allows the user to fit generative models of
mutation from an annotated mutation dataset, and then further to produce tunable linear
estimators of exome-wide biomarkers. It also contains functions to simulate mutation
annotated format (MAF) data, as well as to analyse the output and performance of models.
}
