% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsclc_data.R
\docType{data}
\name{nsclc_maf}
\alias{nsclc_maf}
\title{Non-Small Cell Lung Cancer MAF Data}
\format{
An annotated mutation dataframe with 6 columns and 299855 rows:
\describe{
\item{Tumor_Sample_Barcode}{A sample id for each mutation.}
\item{Hugo_Symbol}{The name of the gene location for each mutation.}
\item{Variant_Classification}{The mutation type for each mutation.}
\item{Chromosome}{Chromosome on which the mutation occurred.}
\item{Start_Position}{Start nucleotide location for mutation.}
\item{End_Position}{End nucleotide location for mutation.}
}
}
\source{
\url{https://www.cbioportal.org/study/summary?id=nsclc_tcga_broad_2016}
}
\usage{
nsclc_maf
}
\description{
A pre-loaded mutation dataset from Campbell et. al (2016), downloaded from The Cancer Genome Atlas.
}
\keyword{datasets}
