% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_tib_tables}
\alias{example_tib_tables}
\title{Tumour Indel Burden of Example Train, Validation and Test Data.}
\format{
A list with threeobjects: 'train', 'val' and 'test'. Each is a dataframe with two columns:
\describe{
\item{Tumor_Sample_Barcode}{A unique ID for each sample.}
\item{TIB}{The value of Tumour Indel Burden for that sample.}
}
}
\usage{
example_tib_tables
}
\description{
An example output produced by using the function get_biomarker_tables(), applied to the example MAF data pre-loaded in example_maf_data$maf.
}
\keyword{datasets}
