% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_model.R
\name{vis_model_fit}
\alias{vis_model_fit}
\title{Visualise Generative Model Fit}
\usage{
vis_model_fit(
  gen_model,
  x_sparsity = FALSE,
  y_sparsity = FALSE,
  mut_type = NULL
)
}
\arguments{
\item{gen_model}{(list)
A generative model fitted by fit_gen_model()}

\item{x_sparsity}{Show model sparsity on x axis rather than lambda.}

\item{y_sparsity}{Show model sparsity on y axis rather than deviance.}

\item{mut_type}{Produce separate plots for each mutation type.}
}
\value{
Summary plot of the generative model fit across folds.
}
\description{
A function to visualise how well a general model has fitted to a mutation dataset across cross-validation folds. Designed to produce a similar output to glmnet's function plot.cv.glmnet.
}
\examples{
p <- vis_model_fit(example_gen_model)

}
